+++ lib/src/ResponseHandler.dart
@@ -1,7 +1,0 @@
-part of route_provider;
-
-abstract class ResponseHandler {
-  ResponseHandler();
-
-  Future response(HttpRequest request, Map vars) async {}
-}