+++ test/src/RouteProviderTest.dart
@@ -6,46 +6,90 @@ import '../TestContext/TestContext.dart';
 void main() {
 
     TestContext context = new TestContext(port: 4040);
-    RouteProvider routeProvider;
+    Router router;
 
     setUpAll(() async {
-        routeProvider = await context.init();
-        routeProvider.start();
+        router = await context.init();
+        router.start();
     });
 
-    setUp(() {
-        routeProvider.clear();
+    setUp(() async {
+        await router.clear();
     });
 
     tearDownAll(() {
-        routeProvider.stop();
+        router.stop();
     });
 
-    test('If the controller in a route return a notFoundError, the server delivers a 404 status', () async {
-        String path = '/test';
-        routeProvider.route(
-            url: path,
-            controller: context.controller.notFoundError,
-            responser: context.responser.none,
-            auth: context.auth.allowed
-        );
-
-        HttpClientResponse response = await context.get(path);
-
-        expect(response.statusCode, equals(HttpStatus.notFound));
+    group('Responsers set ContentType', () {
+        group('JSON-Responser set "application/json"', () {
+            test('normal case', () async {
+                String path = '/test1';
+                router.route(
+                    url: path,
+                    responser: context.responser.json,
+                );
+                HttpClientResponse response = await context.client.get(path);
+                expect(response.headers.contentType.mimeType, equals('application/json'));
+            });
+            test('error from controller', () async {
+                String path = '/test2';
+                router.route(
+                    url: path,
+                    controller: context.controller.notFoundError,
+                    responser: context.responser.json,
+                    auth: context.auth.allowed
+                );
+                HttpClientResponse response = await context.client.get(path);
+                expect(response.headers.contentType.mimeType, equals('application/json'));
+            });
+            test('error from auth', () async {
+                String path = '/test3';
+                router.route(
+                    url: path,
+                    responser: context.responser.json,
+                    auth: context.auth.allowed
+                );
+                HttpClientResponse response = await context.client.get(path);
+                expect(response.headers.contentType.mimeType, equals('application/json'));
+            });
+        });
     });
 
-    test('If the auth in a route do not allow the access, the server delivers a 401 status', () async {
-        String path = '/test';
-        routeProvider.route(
-            url: path,
-            controller: context.controller.notFoundError,
-            responser: context.responser.none,
-            auth: context.auth.denyed
-        );
+    group('Controller errors', () {
+        test('If a controller return a notFoundError, the server delivers a 404 status', () async {
+            String path = '/test4';
+            router.route(
+                url: path,
+                controller: context.controller.notFoundError,
+                responser: context.responser.none,
+                auth: context.auth.allowed
+            );
 
-        HttpClientResponse response = await context.get(path);
+            HttpClientResponse response = await context.client.get(path);
 
-        expect(response.statusCode, equals(HttpStatus.unauthorized));
+            expect(response.statusCode, equals(HttpStatus.notFound));
+            expect(response.contentLength, equals(-1));
+            expect(response.headers.contentType.mimeType, equals('text/plain'));
+        });
+        test('If the auth in a route do not allow the access, the server delivers a 401 status', () async {
+            String path = '/test5';
+            router.route(
+                url: path,
+                controller: context.controller.notFoundError,
+                responser: context.responser.none,
+                auth: context.auth.denyed
+            );
+
+            HttpClientResponse response = await context.client.get(path);
+
+            expect(response.statusCode, equals(HttpStatus.unauthorized));
+            expect(response.contentLength, equals(-1));
+            expect(response.headers.contentType.mimeType, equals('text/plain'));
+        });
     });
+
+
+
+
 }